requirements.txt
aiosmtplib==2.0.2
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
bcrypt==4.1.2
blinker==1.8.2
boto3==1.34.74
botocore==1.34.74
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.41.0
dnspython==2.6.1
ecdsa==0.18.0
email_validator==2.1.1
exceptiongroup==1.2.0
fastapi==0.112.1
fastapi-limiter==0.1.6
fastapi-mail==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
pyasn1==0.6.0
pydantic==2.8.2
pydantic-settings==2.5.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
redis==5.1.0b4
requests==2.31.0
rsa==4.9
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.38.2
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.30.6